Make parameter_type required in ContentPackResourceController
