+++ st2api/st2api/controllers/v1/ruleviews.py
@@ -24,6 +24,7 @@ from st2common.models.system.common import ResourceReference
 from st2common.persistence.action import Action
 from st2common.persistence.rule import Rule
 from st2common.persistence.trigger import TriggerType, Trigger
+from st2common.rbac.types import PermissionType
 
 http_client = six.moves.http_client
 
@@ -93,7 +94,7 @@ class RuleViewController(resource.ContentPackResourceController):
         return rules
 
     def get_one(self, ref_or_id):
-        rule = self._get_one(ref_or_id)
+        rule = self._get_one(ref_or_id, permission_type=PermissionType.RULE_VIEW)
         result = self._append_view_properties([rule.json])[0]
         rule.json = result
         return rule
